requirements.txt
Django==4.2.0
djangorestframework==3.14.0
Pillow==9.5.0
django-filter==23.3